Cleaned up auth route and controller
